requirements.txt
aiohttp
beautifulsoup4>=4.12.0
httpx
fastmcp
requests>=2.31.0
mcp-server>=0.1.0